requirements.txt
cffi==1.15.1
click==8.1.3
Flask==2.2.2
importlib-metadata==6.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyNaCl==1.5.0
Werkzeug==2.2.2
zipp==3.11.0
